django_project/requirements.txt
asgiref==3.2.10
beautifulsoup4==4.10.0
bootstrap4==0.1.0
certifi==2018.10.15
chardet==3.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==3.0.1
django-crispy-forms==1.9.2
django-filter==2.4.0
django-forms-bootstrap==3.1.0
django-formtools==2.2
django-import-export==2.7.1
django-search-views==0.3.7
django-smart-selects==1.6.0
django-user-visit==0.4
django-widget-tweaks==1.4.12
djangorestframework==3.12.2
et-xmlfile==1.1.0
html5lib==1.1
idna==2.7
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==7.0.0
psycopg2==2.9.9
PyPDF2==1.26.0
python-barcode==0.13.1
pytz==2018.5
PyYAML==6.0
reportlab==3.5.51
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.2.0
ua-parser==0.10.0
user-agents==2.2.0
webencodings==0.5.1
xhtml2pdf==0.2.4
xlrd==2.0.1
xlwt==1.3.0
